+++ src/components/index.js
@@ -1,0 +1,8 @@
+import Header from './Header';
+import ItemList from './ItemList';
+import ListTitle from './ListTitle';
+import MoreButton from './MoreButton';
+import movieItem from './movieItem';
+import WholeScreenMessageAlert from './WholeScreenMessageAlert';
+
+export { Header, ItemList, ListTitle, MoreButton, movieItem, WholeScreenMessageAlert };
